Fix optional trailing delimiter on replacements
The problem before was that the second capturing group was getting too
greedy and taking the whole of the middle of the string, so for example
's/test/blah/' was turning into 'test/blah' being the matched string,
which doesn't usually match any previous text. Instead, now the second
and third capturing groups are lazy, so that the optional delimiter at
the end can actually function correctly.