regex: make a trailing separator actually terminate the command
As it stands, the bot interprets "s/foo/bar/ blah blah" as a request to
replace "foo" with "bar/ blah blah", which is exceedingly unlikely to be
the behavior the user desires.  Instead, treat a trailing separator
character as the end of the substitution command and ignore anything
else that follows in the message.