weather: convert current temperature instead of bounds and avoid rounding
At the moment, due to the use of int() to round down the various
temperature bounds in icon() and color(), the temperature boundaries
between colors and icons depend on choice of unit.  That leads to fun
like this:

    <sluo> create: weather
    <create> Berkeley: 74.57°F ☀ clear sky; wind: 6.93m/s 320° (NW); humidity: 49; UV index: 6.26
    <sluo> create: weather -c
    <create> Berkeley: 23.65°C ☢ clear sky; wind: 3.1m/s 320° (NW); humidity: 49; UV index: 6.26

Don't round temperatures to avoid this issue; while we're at it, convert
the current temperature instead of the bounds, to avoid doing a bunch of
conversions every time we call color().